requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-pptx==0.6.23
Pillow==10.1.0
werkzeug==3.0.1
gunicorn==21.2.0
